Store and restore the router from json (#6)
The routes in the router usually never change during runtime. We could
construct the router during compile time in some cases. This feature
allows us to save and restore the configuration of the router so we can
initialize the router faster and more convenient!

Also the typescript configuration was fixed, it was not generating types
and sourcemaps, that had quite a few consequences!